commit UpdateProduct  with updateMethod exception tests 4

+++ src/main/java/com/gfttraining/productAPI/controllers/ProductController.java
@@ -63,7 +63,7 @@ public class ProductController {
     }
 
     @PutMapping("/products/{id_product}")
-    public ResponseEntity<Product> putUpdate (@PathVariable int id_product,  @RequestBody @Valid ProductRequest updateProductRequest) throws NonExistingProductException {
+    public ResponseEntity<Product> putUpdate (@PathVariable Long id_product,  @RequestBody @Valid ProductRequest updateProductRequest) throws NonExistingProductException {
         return new ResponseEntity<>(
             productService.updateProduct(id_product, updateProductRequest),
             HttpStatus.OK
